Merge pull request #68 from RustLangES/only-new-async
Rust 2024

+++ listings/ch11-writing-automated-tests/listing-11-03/src/lib.rs
@@ -1,5 +1,4 @@
-// ANCHOR: here
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: u64, right: u64) -> u64 {
     left + right
 }
 
@@ -12,10 +11,9 @@ mod tests {
         let result = add(2, 2);
         assert_eq!(result, 4);
     }
-    
+
     #[test]
     fn another() {
         panic!("Make this test fail");
     }
 }
-// ANCHOR_END: here
